+++ src/raytracer/ray.c
@@ -16,7 +16,7 @@ t_ray			ray_transform(t_ray ray, t_matrix transformation_matrix)
 	);
 }
 
-t_tuple	ft_position(t_ray r, float t)
+t_tuple	ft_position(t_ray r, double t)
 {
 	return add_tuple(r.origin, multiply_tuple(r.direction, t));
 }
